@@ -1,36 +1,19 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>. */
-
 package collector
 
 import (
-	"io/ioutil"
+	"github.com/stretchr/testify/assert"
+	"io"
 	"os"
 	"testing"
 )
 
 func TestCPUsMetrics(t *testing.T) {
-	// Read the input data from a file
-	file, err := os.Open("test_data/sinfo_cpus.txt")
-	if err != nil {
-		t.Fatalf("Can not open test data: %v", err)
-	}
-	data, err := ioutil.ReadAll(file)
-	t.Logf("%+v", ParseCPUsMetrics(data))
-}
+	file, _ := os.Open("fixtures/sinfo/cpus.txt")
+	data, _ := io.ReadAll(file)
+	cpus := ParseCPUsMetrics(data)
 
-func TestCPUssGetMetrics(t *testing.T) {
-	t.Logf("%+v", CPUsGetMetrics())
+	assert.Equal(t, 5725.0, cpus.alloc, "Miscount of alloc CPUs")
+	assert.Equal(t, 877.0, cpus.idle, "Miscount of idle CPUs")
+	assert.Equal(t, 34.0, cpus.other, "Miscount of other CPUs")
+	assert.Equal(t, 6636.0, cpus.total, "Miscount of total CPUs")
 }
