@@ -1,36 +1,27 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>. */
-
 package collector
 
 import (
-	"io/ioutil"
+	"github.com/stretchr/testify/assert"
+	"io"
 	"os"
 	"testing"
 )
 
 func TestNodesMetrics(t *testing.T) {
 	// Read the input data from a file
-	file, err := os.Open("test_data/sinfo.txt")
-	if err != nil {
-		t.Fatalf("Can not open test data: %v", err)
-	}
-	data, err := ioutil.ReadAll(file)
-	t.Logf("%+v", ParseNodesMetrics(data))
-}
+	file, _ := os.Open("fixtures/sinfo/nodes.txt")
+	data, _ := io.ReadAll(file)
+	nodes := ParseNodesMetrics(data)
 
-func TestNodesGetMetrics(t *testing.T) {
-	t.Logf("%+v", NodesGetMetrics())
+	assert.Equal(t, 1.0, nodes.alloc)
+	assert.Equal(t, 0.0, nodes.comp)
+	assert.Equal(t, 0.0, nodes.down)
+	assert.Equal(t, 7.0, nodes.drain)
+	assert.Equal(t, 0.0, nodes.err)
+	assert.Equal(t, 0.0, nodes.fail)
+	assert.Equal(t, 9.0, nodes.idle)
+	assert.Equal(t, 0.0, nodes.maint)
+	assert.Equal(t, 20.0, nodes.mix)
+	assert.Equal(t, 0.0, nodes.resv)
+	assert.Equal(t, 0.0, nodes.plnd)
 }
