@@ -1,36 +1,29 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>. */
-
 package collector
 
 import (
-	"io/ioutil"
+	"github.com/stretchr/testify/assert"
+	"io"
 	"os"
 	"testing"
 )
 
 func TestParseQueueMetrics(t *testing.T) {
 	// Read the input data from a file
-	file, err := os.Open("test_data/squeue.txt")
-	if err != nil {
-		t.Fatalf("Can not open test data: %v", err)
-	}
-	data, err := ioutil.ReadAll(file)
-	t.Logf("%+v", ParseQueueMetrics(data))
-}
+	file, _ := os.Open("fixtures/squeue/queue.txt")
+	data, _ := io.ReadAll(file)
+	queueMetrics := ParseQueueMetrics(data)
 
-func TestQueueGetMetrics(t *testing.T) {
-	t.Logf("%+v", QueueGetMetrics())
+	assert.Equal(t, 32.0, queueMetrics.pending, "Miscount of pending jobs")
+	assert.Equal(t, 0.0, queueMetrics.pendingDep, "Miscount of pendingDep jobs")
+	assert.Equal(t, 27.0, queueMetrics.running, "Miscount of running jobs")
+	assert.Equal(t, 0.0, queueMetrics.suspended, "Miscount of suspended jobs")
+	assert.Equal(t, 1.0, queueMetrics.cancelled, "Miscount of cancelled jobs")
+	assert.Equal(t, 1.0, queueMetrics.completing, "Miscount of completing jobs")
+	assert.Equal(t, 1.0, queueMetrics.completed, "Miscount of completed jobs")
+	assert.Equal(t, 0.0, queueMetrics.configuring, "Miscount of configuring jobs")
+	assert.Equal(t, 0.0, queueMetrics.failed, "Miscount of failed jobs")
+	assert.Equal(t, 0.0, queueMetrics.timeout, "Miscount of timeout jobs")
+	assert.Equal(t, 0.0, queueMetrics.preempted, "Miscount of preempted jobs")
+	assert.Equal(t, 0.0, queueMetrics.nodeFail, "Miscount of nodeFail jobs")
+	assert.Equal(t, 0.0, queueMetrics.outOfMemory, "Miscount of outOfMemory jobs")
 }
