@@ -1,4 +1,6 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
+/*
+	Copyright 2017 Victor Penso, Matteo Dessalvi
+	Copyright 2024 Oleh Astappiev
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,87 +18,70 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package collector
 
 import (
-	"github.com/prometheus/client_golang/prometheus"
-	"io/ioutil"
-	"log"
-	"os/exec"
 	"strconv"
 	"strings"
+
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
-type CPUsMetrics struct {
+type CPUs struct {
 	alloc float64
 	idle  float64
 	other float64
 	total float64
 }
 
-func CPUsGetMetrics() *CPUsMetrics {
-	return ParseCPUsMetrics(CPUsData())
+func ParseCPUs(input string) CPUs {
+	var cm CPUs
+	if strings.Contains(input, "/") {
+		parts := strings.Split(strings.TrimSpace(input), "/")
+		cm.alloc, _ = strconv.ParseFloat(parts[0], 64)
+		cm.idle, _ = strconv.ParseFloat(parts[1], 64)
+		cm.other, _ = strconv.ParseFloat(parts[2], 64)
+		cm.total, _ = strconv.ParseFloat(parts[3], 64)
+	}
+	return cm
 }
 
-func ParseCPUsMetrics(input []byte) *CPUsMetrics {
-	var cm CPUsMetrics
-	if strings.Contains(string(input), "/") {
-		splitted := strings.Split(strings.TrimSpace(string(input)), "/")
-		cm.alloc, _ = strconv.ParseFloat(splitted[0], 64)
-		cm.idle, _ = strconv.ParseFloat(splitted[1], 64)
-		cm.other, _ = strconv.ParseFloat(splitted[2], 64)
-		cm.total, _ = strconv.ParseFloat(splitted[3], 64)
-	}
-	return &cm
-}
-
-// Execute the sinfo command and return its output
-func CPUsData() []byte {
-	cmd := exec.Command("sinfo", "-h", "-a", "-o %C")
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := cmd.Start(); err != nil {
-		log.Fatal(err)
-	}
-	out, _ := ioutil.ReadAll(stdout)
-	if err := cmd.Wait(); err != nil {
-		log.Fatal(err)
-	}
-	return out
-}
-
-/*
- * Implement the Prometheus Collector interface and feed the
- * Slurm scheduler metrics into it.
- * https://godoc.org/github.com/prometheus/client_golang/prometheus#Collector
- */
-
-func NewCPUsCollector() *CPUsCollector {
-	return &CPUsCollector{
-		alloc: prometheus.NewDesc("slurm_cpus_alloc", "Allocated CPUs", nil, nil),
-		idle:  prometheus.NewDesc("slurm_cpus_idle", "Idle CPUs", nil, nil),
-		other: prometheus.NewDesc("slurm_cpus_other", "Mix CPUs", nil, nil),
-		total: prometheus.NewDesc("slurm_cpus_total", "Total CPUs", nil, nil),
-	}
+func ParseCPUsMetrics(input []byte) *CPUs {
+	cpu := ParseCPUs(string(input))
+	return &cpu
 }
 
 type CPUsCollector struct {
-	alloc *prometheus.Desc
-	idle  *prometheus.Desc
-	other *prometheus.Desc
-	total *prometheus.Desc
+	alloc  *prometheus.Desc
+	idle   *prometheus.Desc
+	other  *prometheus.Desc
+	total  *prometheus.Desc
+	logger log.Logger
 }
 
-// Send all metric descriptions
-func (cc *CPUsCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- cc.alloc
-	ch <- cc.idle
-	ch <- cc.other
-	ch <- cc.total
+func init() {
+	registerCollector("cpus", defaultEnabled, NewCPUsCollector)
 }
-func (cc *CPUsCollector) Collect(ch chan<- prometheus.Metric) {
-	cm := CPUsGetMetrics()
+
+func NewCPUsCollector(logger log.Logger) (Collector, error) {
+	return &CPUsCollector{
+		logger: logger,
+		alloc:  prometheus.NewDesc("slurm_cpus_alloc", "Allocated CPUs", nil, nil),
+		idle:   prometheus.NewDesc("slurm_cpus_idle", "Idle CPUs", nil, nil),
+		other:  prometheus.NewDesc("slurm_cpus_other", "Mix CPUs", nil, nil),
+		total:  prometheus.NewDesc("slurm_cpus_total", "Total CPUs", nil, nil),
+	}, nil
+}
+
+func (cc *CPUsCollector) Collect(ch chan<- prometheus.Metric) error {
+	out, err := RunCommand("sinfo", "-h", "-a", "-o %C")
+	if err != nil {
+		return err
+	}
+
+	cm := ParseCPUsMetrics(out)
 	ch <- prometheus.MustNewConstMetric(cc.alloc, prometheus.GaugeValue, cm.alloc)
 	ch <- prometheus.MustNewConstMetric(cc.idle, prometheus.GaugeValue, cm.idle)
 	ch <- prometheus.MustNewConstMetric(cc.other, prometheus.GaugeValue, cm.other)
 	ch <- prometheus.MustNewConstMetric(cc.total, prometheus.GaugeValue, cm.total)
+
+	return nil
 }
