@@ -1,4 +1,6 @@
-/* Copyright 2021 Chris Read
+/*
+	Copyright 2021 Chris Read
+	Copyright 2024 Oleh Astappiev
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,39 +18,54 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package collector
 
 import (
-	"log"
-	"os/exec"
 	"sort"
 	"strconv"
 	"strings"
 
+	"github.com/go-kit/log"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
 // NodeMetrics stores metrics for each node
 type NodeMetrics struct {
-	memAlloc   uint64
-	memTotal   uint64
-	cpuAlloc   uint64
-	cpuIdle    uint64
-	cpuOther   uint64
-	cpuTotal   uint64
-	gpuAlloc   uint64
-	gpuTotal   uint64
-	gpuType    string
+	memAlloc   float64
+	memTotal   float64
+	cpu        CPUs
+	tres       []TRES
+	tresUsed   []TRES
 	nodeStatus string
 }
 
-func NodeGetMetrics() map[string]*NodeMetrics {
-	return ParseNodeMetrics(NodeData())
+type TRES struct {
+	kind  string
+	name  string
+	count float64
+}
+
+func ParseTres(input string) []TRES {
+	results := make([]TRES, 0)
+
+	// remove parenthesis and all content inside
+	for strings.Contains(input, "(") {
+		openIndex := strings.Index(input, "(")
+		closeIndex := strings.Index(input, ")")
+		input = input[:openIndex] + input[closeIndex+1:]
+	}
+
+	for _, single := range strings.Split(input, ",") {
+		tresInfo := strings.Split(single, ":")
+		tresCount, _ := strconv.ParseFloat(tresInfo[2], 64)
+		results = append(results, TRES{kind: tresInfo[0], name: tresInfo[1], count: tresCount})
+	}
+
+	return results
 }
 
 // ParseNodeMetrics takes the output of sinfo with node data
-// It returns a map of metrics per node
 func ParseNodeMetrics(input []byte) map[string]*NodeMetrics {
 	nodes := make(map[string]*NodeMetrics)
-	lines := strings.Split(string(input), "\n")
 
+	lines := SplitLines(input)
 	// Sort and remove all the duplicates from the 'sinfo' output
 	sort.Strings(lines)
 	linesUniq := RemoveDuplicates(lines)
@@ -56,53 +73,25 @@ func ParseNodeMetrics(input []byte) map[string]*NodeMetrics {
 	for _, line := range linesUniq {
 		node := strings.Fields(line)
 		nodeName := node[0]
-		nodeStatus := node[5] // mixed, allocated, etc.
+		nodeStatus := node[4] // mixed, allocated, etc.
 
-		nodes[nodeName] = &NodeMetrics{0, 0, 0, 0, 0, 0, 0, 0, "", ""}
+		nodes[nodeName] = &NodeMetrics{}
 
-		memAlloc, _ := strconv.ParseUint(node[1], 10, 64)
-		memTotal, _ := strconv.ParseUint(node[2], 10, 64)
+		nodes[nodeName].cpu = ParseCPUs(node[3])
+		memAlloc, _ := strconv.ParseFloat(node[1], 64)
+		memTotal, _ := strconv.ParseFloat(node[2], 64)
 
-		cpuInfo := strings.Split(node[3], "/")
-		cpuAlloc, _ := strconv.ParseUint(cpuInfo[0], 10, 64)
-		cpuIdle, _ := strconv.ParseUint(cpuInfo[1], 10, 64)
-		cpuOther, _ := strconv.ParseUint(cpuInfo[2], 10, 64)
-		cpuTotal, _ := strconv.ParseUint(cpuInfo[3], 10, 64)
-
-		if node[5] != "(null)" {
-			// Ignore everything after opening parenthesis and split into type, name and count
-			availableTRES := strings.Split(strings.Split(node[5], "(")[0], ":")
-			usedTRES := strings.Split(strings.Split(node[6], "(")[0], ":")
-			gpuType := availableTRES[1]
-			gpuTotal, _ := strconv.ParseUint(availableTRES[2], 10, 64)
-			gpuAlloc, _ := strconv.ParseUint(usedTRES[2], 10, 64)
-
-			nodes[nodeName].gpuAlloc = gpuAlloc
-			nodes[nodeName].gpuTotal = gpuTotal
-			nodes[nodeName].gpuType = gpuType
+		if len(node) >= 6 && node[5] != "(null)" && len(node[5]) > 0 {
+			nodes[nodeName].tres = ParseTres(node[5])
+			nodes[nodeName].tresUsed = ParseTres(node[6])
 		}
 
 		nodes[nodeName].memAlloc = memAlloc
 		nodes[nodeName].memTotal = memTotal
-		nodes[nodeName].cpuAlloc = cpuAlloc
-		nodes[nodeName].cpuIdle = cpuIdle
-		nodes[nodeName].cpuOther = cpuOther
-		nodes[nodeName].cpuTotal = cpuTotal
 		nodes[nodeName].nodeStatus = nodeStatus
 	}
 
 	return nodes
-}
-
-// NodeData executes the sinfo command to get data for each node
-// It returns the output of the sinfo command
-func NodeData() []byte {
-	cmd := exec.Command("sinfo", "-h", "-a", "-N", "-O", "NodeList: ,AllocMem: ,Memory: ,CPUsState: ,StateLong: ,Gres: ,Gresused:")
-	out, err := cmd.Output()
-	if err != nil {
-		log.Fatal(err)
-	}
-	return out
 }
 
 type NodeCollector struct {
@@ -114,50 +103,54 @@ type NodeCollector struct {
 	memTotal *prometheus.Desc
 	gpuAlloc *prometheus.Desc
 	gpuTotal *prometheus.Desc
+	logger   log.Logger
+}
+
+func init() {
+	registerCollector("node", defaultEnabled, NewNodeCollector)
 }
 
 // NewNodeCollector creates a Prometheus collector to keep all our stats in
 // It returns a set of collections for consumption
-func NewNodeCollector() *NodeCollector {
-	labels := []string{"node", "status"}
-	labels_gpu := []string{"node", "status", "gputype"}
-
+func NewNodeCollector(logger log.Logger) (Collector, error) {
 	return &NodeCollector{
-		cpuAlloc: prometheus.NewDesc("slurm_node_cpu_alloc", "Allocated CPUs per node", labels, nil),
-		cpuIdle:  prometheus.NewDesc("slurm_node_cpu_idle", "Idle CPUs per node", labels, nil),
-		cpuOther: prometheus.NewDesc("slurm_node_cpu_other", "Other CPUs per node", labels, nil),
-		cpuTotal: prometheus.NewDesc("slurm_node_cpu_total", "Total CPUs per node", labels, nil),
-		memAlloc: prometheus.NewDesc("slurm_node_mem_alloc", "Allocated memory per node", labels, nil),
-		memTotal: prometheus.NewDesc("slurm_node_mem_total", "Total memory per node", labels, nil),
-		gpuAlloc: prometheus.NewDesc("slurm_node_gpu_alloc", "Allocated GPUs per node", labels_gpu, nil),
-		gpuTotal: prometheus.NewDesc("slurm_node_gpu_total", "Total GPUs per node", labels_gpu, nil),
+		logger:   logger,
+		cpuAlloc: prometheus.NewDesc("slurm_node_cpu_alloc", "Allocated CPUs per node", []string{"node", "status"}, nil),
+		cpuIdle:  prometheus.NewDesc("slurm_node_cpu_idle", "Idle CPUs per node", []string{"node", "status"}, nil),
+		cpuOther: prometheus.NewDesc("slurm_node_cpu_other", "Other CPUs per node", []string{"node", "status"}, nil),
+		cpuTotal: prometheus.NewDesc("slurm_node_cpu_total", "Total CPUs per node", []string{"node", "status"}, nil),
+		memAlloc: prometheus.NewDesc("slurm_node_mem_alloc", "Allocated memory per node", []string{"node", "status"}, nil),
+		memTotal: prometheus.NewDesc("slurm_node_mem_total", "Total memory per node", []string{"node", "status"}, nil),
+		gpuAlloc: prometheus.NewDesc("slurm_node_gpu_alloc", "Allocated GPUs per node", []string{"node", "status", "gputype"}, nil),
+		gpuTotal: prometheus.NewDesc("slurm_node_gpu_total", "Total GPUs per node", []string{"node", "status", "gputype"}, nil),
+	}, nil
+}
+
+func (c *NodeCollector) Collect(ch chan<- prometheus.Metric) error {
+	out, err := RunCommand("sinfo", "-h", "-a", "-N", "-O", "NodeList: ,AllocMem: ,Memory: ,CPUsState: ,StateLong: ,Gres: ,Gresused:")
+	if err != nil {
+		return err
 	}
-}
 
-// Send all metric descriptions
-func (nc *NodeCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- nc.cpuAlloc
-	ch <- nc.cpuIdle
-	ch <- nc.cpuOther
-	ch <- nc.cpuTotal
-	ch <- nc.memAlloc
-	ch <- nc.memTotal
-	ch <- nc.gpuAlloc
-	ch <- nc.gpuTotal
-}
-
-func (nc *NodeCollector) Collect(ch chan<- prometheus.Metric) {
-	nodes := NodeGetMetrics()
+	nodes := ParseNodeMetrics(out)
 	for node := range nodes {
-		ch <- prometheus.MustNewConstMetric(nc.cpuAlloc, prometheus.GaugeValue, float64(nodes[node].cpuAlloc), node, nodes[node].nodeStatus)
-		ch <- prometheus.MustNewConstMetric(nc.cpuIdle, prometheus.GaugeValue, float64(nodes[node].cpuIdle), node, nodes[node].nodeStatus)
-		ch <- prometheus.MustNewConstMetric(nc.cpuOther, prometheus.GaugeValue, float64(nodes[node].cpuOther), node, nodes[node].nodeStatus)
-		ch <- prometheus.MustNewConstMetric(nc.cpuTotal, prometheus.GaugeValue, float64(nodes[node].cpuTotal), node, nodes[node].nodeStatus)
-		ch <- prometheus.MustNewConstMetric(nc.memAlloc, prometheus.GaugeValue, float64(nodes[node].memAlloc), node, nodes[node].nodeStatus)
-		ch <- prometheus.MustNewConstMetric(nc.memTotal, prometheus.GaugeValue, float64(nodes[node].memTotal), node, nodes[node].nodeStatus)
-		if nodes[node].gpuType != "" {
-			ch <- prometheus.MustNewConstMetric(nc.gpuAlloc, prometheus.GaugeValue, float64(nodes[node].gpuAlloc), node, nodes[node].nodeStatus, nodes[node].gpuType)
-			ch <- prometheus.MustNewConstMetric(nc.gpuTotal, prometheus.GaugeValue, float64(nodes[node].gpuTotal), node, nodes[node].nodeStatus, nodes[node].gpuType)
+		ch <- prometheus.MustNewConstMetric(c.cpuAlloc, prometheus.GaugeValue, nodes[node].cpu.alloc, node, nodes[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(c.cpuIdle, prometheus.GaugeValue, nodes[node].cpu.idle, node, nodes[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(c.cpuOther, prometheus.GaugeValue, nodes[node].cpu.other, node, nodes[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(c.cpuTotal, prometheus.GaugeValue, nodes[node].cpu.total, node, nodes[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(c.memAlloc, prometheus.GaugeValue, nodes[node].memAlloc, node, nodes[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(c.memTotal, prometheus.GaugeValue, nodes[node].memTotal, node, nodes[node].nodeStatus)
+		if nodes[node].tresUsed != nil && len(nodes[node].tresUsed) != 0 {
+			for _, tres := range nodes[node].tresUsed {
+				ch <- prometheus.MustNewConstMetric(c.gpuAlloc, prometheus.GaugeValue, tres.count, node, nodes[node].nodeStatus, tres.name)
+			}
+		}
+		if nodes[node].tres != nil && len(nodes[node].tres) != 0 {
+			for _, tres := range nodes[node].tres {
+				ch <- prometheus.MustNewConstMetric(c.gpuTotal, prometheus.GaugeValue, tres.count, node, nodes[node].nodeStatus, tres.name)
+			}
 		}
 	}
+
+	return nil
 }
