@@ -1,4 +1,6 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
+/*
+	Copyright 2017 Victor Penso, Matteo Dessalvi
+	Copyright 2024 Oleh Astappiev
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,14 +18,13 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package collector
 
 import (
-	"github.com/prometheus/client_golang/prometheus"
-	"io/ioutil"
-	"log"
-	"os/exec"
 	"regexp"
 	"sort"
 	"strconv"
 	"strings"
+
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
 type NodesMetrics struct {
@@ -40,51 +41,34 @@ type NodesMetrics struct {
 	plnd  float64
 }
 
-func NodesGetMetrics() *NodesMetrics {
-	return ParseNodesMetrics(NodesData())
-}
-
-func RemoveDuplicates(s []string) []string {
-	m := map[string]bool{}
-	t := []string{}
-
-	// Walk through the slice 's' and for each value we haven't seen so far, append it to 't'.
-	for _, v := range s {
-		if _, seen := m[v]; !seen {
-			if len(v) > 0 {
-				t = append(t, v)
-				m[v] = true
-			}
-		}
-	}
-
-	return t
-}
-
 func ParseNodesMetrics(input []byte) *NodesMetrics {
-	var nm NodesMetrics
-	lines := strings.Split(string(input), "\n")
+	var (
+		alloc = regexp.MustCompile(`^alloc`)
+		comp  = regexp.MustCompile(`^comp`)
+		down  = regexp.MustCompile(`^down`)
+		drain = regexp.MustCompile(`^drain`)
+		fail  = regexp.MustCompile(`^fail`)
+		err   = regexp.MustCompile(`^err`)
+		idle  = regexp.MustCompile(`^idle`)
+		maint = regexp.MustCompile(`^maint`)
+		mix   = regexp.MustCompile(`^mix`)
+		resv  = regexp.MustCompile(`^res`)
+		plnd  = regexp.MustCompile(`^plan`)
+	)
 
+	lines := SplitLines(input)
 	// Sort and remove all the duplicates from the 'sinfo' output
 	sort.Strings(lines)
-	lines_uniq := RemoveDuplicates(lines)
+	linesUniq := RemoveDuplicates(lines)
 
-	for _, line := range lines_uniq {
+	var nm NodesMetrics
+	for _, line := range linesUniq {
 		if strings.Contains(line, ",") {
-			split := strings.Split(line, ",")
-			count, _ := strconv.ParseFloat(strings.TrimSpace(split[0]), 64)
-			state := split[1]
-			alloc := regexp.MustCompile(`^alloc`)
-			comp := regexp.MustCompile(`^comp`)
-			down := regexp.MustCompile(`^down`)
-			drain := regexp.MustCompile(`^drain`)
-			fail := regexp.MustCompile(`^fail`)
-			err := regexp.MustCompile(`^err`)
-			idle := regexp.MustCompile(`^idle`)
-			maint := regexp.MustCompile(`^maint`)
-			mix := regexp.MustCompile(`^mix`)
-			resv := regexp.MustCompile(`^res`)
-			plnd := regexp.MustCompile(`^plan`)
+			parts := strings.Split(line, ",")
+
+			count, _ := strconv.ParseFloat(strings.TrimSpace(parts[0]), 64)
+			state := parts[1]
+
 			switch {
 			case alloc.MatchString(state) == true:
 				nm.alloc += count
@@ -114,75 +98,49 @@ func ParseNodesMetrics(input []byte) *NodesMetrics {
 	return &nm
 }
 
-// Execute the sinfo command and return its output
-func NodesData() []byte {
-	cmd := exec.Command("sinfo", "-h", "-a", "-o %D,%T")
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := cmd.Start(); err != nil {
-		log.Fatal(err)
-	}
-	out, _ := ioutil.ReadAll(stdout)
-	if err := cmd.Wait(); err != nil {
-		log.Fatal(err)
-	}
-	return out
-}
-
-/*
- * Implement the Prometheus Collector interface and feed the
- * Slurm scheduler metrics into it.
- * https://godoc.org/github.com/prometheus/client_golang/prometheus#Collector
- */
-
-func NewNodesCollector() *NodesCollector {
-	return &NodesCollector{
-		alloc: prometheus.NewDesc("slurm_nodes_alloc", "Allocated nodes", nil, nil),
-		comp:  prometheus.NewDesc("slurm_nodes_comp", "Completing nodes", nil, nil),
-		down:  prometheus.NewDesc("slurm_nodes_down", "Down nodes", nil, nil),
-		drain: prometheus.NewDesc("slurm_nodes_drain", "Drain nodes", nil, nil),
-		err:   prometheus.NewDesc("slurm_nodes_err", "Error nodes", nil, nil),
-		fail:  prometheus.NewDesc("slurm_nodes_fail", "Fail nodes", nil, nil),
-		idle:  prometheus.NewDesc("slurm_nodes_idle", "Idle nodes", nil, nil),
-		maint: prometheus.NewDesc("slurm_nodes_maint", "Maint nodes", nil, nil),
-		mix:   prometheus.NewDesc("slurm_nodes_mix", "Mix nodes", nil, nil),
-		resv:  prometheus.NewDesc("slurm_nodes_resv", "Reserved nodes", nil, nil),
-		plnd:  prometheus.NewDesc("slurm_nodes_plnd", "Planned nodes", nil, nil),
-	}
-}
-
 type NodesCollector struct {
-	alloc *prometheus.Desc
-	comp  *prometheus.Desc
-	down  *prometheus.Desc
-	drain *prometheus.Desc
-	err   *prometheus.Desc
-	fail  *prometheus.Desc
-	idle  *prometheus.Desc
-	maint *prometheus.Desc
-	mix   *prometheus.Desc
-	resv  *prometheus.Desc
-	plnd  *prometheus.Desc
+	alloc  *prometheus.Desc
+	comp   *prometheus.Desc
+	down   *prometheus.Desc
+	drain  *prometheus.Desc
+	err    *prometheus.Desc
+	fail   *prometheus.Desc
+	idle   *prometheus.Desc
+	maint  *prometheus.Desc
+	mix    *prometheus.Desc
+	resv   *prometheus.Desc
+	plnd   *prometheus.Desc
+	logger log.Logger
 }
 
-// Send all metric descriptions
-func (nc *NodesCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- nc.alloc
-	ch <- nc.comp
-	ch <- nc.down
-	ch <- nc.drain
-	ch <- nc.err
-	ch <- nc.fail
-	ch <- nc.idle
-	ch <- nc.maint
-	ch <- nc.mix
-	ch <- nc.resv
-	ch <- nc.plnd
+func init() {
+	registerCollector("nodes", defaultEnabled, NewNodesCollector)
 }
-func (nc *NodesCollector) Collect(ch chan<- prometheus.Metric) {
-	nm := NodesGetMetrics()
+
+func NewNodesCollector(logger log.Logger) (Collector, error) {
+	return &NodesCollector{
+		logger: logger,
+		alloc:  prometheus.NewDesc("slurm_nodes_alloc", "Allocated nodes", nil, nil),
+		comp:   prometheus.NewDesc("slurm_nodes_comp", "Completing nodes", nil, nil),
+		down:   prometheus.NewDesc("slurm_nodes_down", "Down nodes", nil, nil),
+		drain:  prometheus.NewDesc("slurm_nodes_drain", "Drain nodes", nil, nil),
+		err:    prometheus.NewDesc("slurm_nodes_err", "Error nodes", nil, nil),
+		fail:   prometheus.NewDesc("slurm_nodes_fail", "Fail nodes", nil, nil),
+		idle:   prometheus.NewDesc("slurm_nodes_idle", "Idle nodes", nil, nil),
+		maint:  prometheus.NewDesc("slurm_nodes_maint", "Maint nodes", nil, nil),
+		mix:    prometheus.NewDesc("slurm_nodes_mix", "Mix nodes", nil, nil),
+		resv:   prometheus.NewDesc("slurm_nodes_resv", "Reserved nodes", nil, nil),
+		plnd:   prometheus.NewDesc("slurm_nodes_plnd", "Planned nodes", nil, nil),
+	}, nil
+}
+
+func (nc *NodesCollector) Collect(ch chan<- prometheus.Metric) error {
+	out, err := RunCommand("sinfo", "-h", "-a", "-o %D,%T")
+	if err != nil {
+		return err
+	}
+
+	nm := ParseNodesMetrics(out)
 	ch <- prometheus.MustNewConstMetric(nc.alloc, prometheus.GaugeValue, nm.alloc)
 	ch <- prometheus.MustNewConstMetric(nc.comp, prometheus.GaugeValue, nm.comp)
 	ch <- prometheus.MustNewConstMetric(nc.down, prometheus.GaugeValue, nm.down)
@@ -194,4 +152,6 @@ func (nc *NodesCollector) Collect(ch chan<- prometheus.Metric) {
 	ch <- prometheus.MustNewConstMetric(nc.mix, prometheus.GaugeValue, nm.mix)
 	ch <- prometheus.MustNewConstMetric(nc.resv, prometheus.GaugeValue, nm.resv)
 	ch <- prometheus.MustNewConstMetric(nc.plnd, prometheus.GaugeValue, nm.plnd)
+
+	return nil
 }
