@@ -1,4 +1,6 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
+/*
+	Copyright 2017 Victor Penso, Matteo Dessalvi
+	Copyright 2024 Oleh Astappiev
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,46 +18,41 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package collector
 
 import (
-	"github.com/prometheus/client_golang/prometheus"
-	"io/ioutil"
-	"log"
-	"os/exec"
 	"strings"
+
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
 type QueueMetrics struct {
-	pending       float64
-	pending_dep   float64
-	running       float64
-	suspended     float64
-	cancelled     float64
-	completing    float64
-	completed     float64
-	configuring   float64
-	failed        float64
-	timeout       float64
-	preempted     float64
-	node_fail     float64
-	out_of_memory float64
-}
-
-// Returns the scheduler metrics
-func QueueGetMetrics() *QueueMetrics {
-	return ParseQueueMetrics(QueueData())
+	pending     float64
+	pendingDep  float64
+	running     float64
+	suspended   float64
+	cancelled   float64
+	completing  float64
+	completed   float64
+	configuring float64
+	failed      float64
+	timeout     float64
+	preempted   float64
+	nodeFail    float64
+	outOfMemory float64
 }
 
 func ParseQueueMetrics(input []byte) *QueueMetrics {
 	var qm QueueMetrics
-	lines := strings.Split(string(input), "\n")
-	for _, line := range lines {
+
+	for _, line := range SplitLines(input) {
 		if strings.Contains(line, ",") {
-			splitted := strings.Split(line, ",")
-			state := splitted[1]
+			parts := strings.Split(line, ",")
+
+			state := parts[1]
 			switch state {
 			case "PENDING":
 				qm.pending++
-				if len(splitted) > 2 && splitted[2] == "Dependency" {
-					qm.pending_dep++
+				if len(parts) > 2 && parts[2] == "Dependency" {
+					qm.pendingDep++
 				}
 			case "RUNNING":
 				qm.running++
@@ -76,92 +73,64 @@ func ParseQueueMetrics(input []byte) *QueueMetrics {
 			case "PREEMPTED":
 				qm.preempted++
 			case "NODE_FAIL":
-				qm.node_fail++
+				qm.nodeFail++
 			case "OUT_OF_MEMORY":
-				qm.out_of_memory++
+				qm.outOfMemory++
 			}
 		}
 	}
 	return &qm
 }
 
-// Execute the squeue command and return its output
-func QueueData() []byte {
-	cmd := exec.Command("squeue", "-a", "-r", "-h", "-o %A,%T,%r", "--states=all")
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := cmd.Start(); err != nil {
-		log.Fatal(err)
-	}
-	out, _ := ioutil.ReadAll(stdout)
-	if err := cmd.Wait(); err != nil {
-		log.Fatal(err)
-	}
-	return out
-}
-
-/*
- * Implement the Prometheus Collector interface and feed the
- * Slurm queue metrics into it.
- * https://godoc.org/github.com/prometheus/client_golang/prometheus#Collector
- */
-
-func NewQueueCollector() *QueueCollector {
-	return &QueueCollector{
-		pending:       prometheus.NewDesc("slurm_queue_pending", "Pending jobs in queue", nil, nil),
-		pending_dep:   prometheus.NewDesc("slurm_queue_pending_dependency", "Pending jobs because of dependency in queue", nil, nil),
-		running:       prometheus.NewDesc("slurm_queue_running", "Running jobs in the cluster", nil, nil),
-		suspended:     prometheus.NewDesc("slurm_queue_suspended", "Suspended jobs in the cluster", nil, nil),
-		cancelled:     prometheus.NewDesc("slurm_queue_cancelled", "Cancelled jobs in the cluster", nil, nil),
-		completing:    prometheus.NewDesc("slurm_queue_completing", "Completing jobs in the cluster", nil, nil),
-		completed:     prometheus.NewDesc("slurm_queue_completed", "Completed jobs in the cluster", nil, nil),
-		configuring:   prometheus.NewDesc("slurm_queue_configuring", "Configuring jobs in the cluster", nil, nil),
-		failed:        prometheus.NewDesc("slurm_queue_failed", "Number of failed jobs", nil, nil),
-		timeout:       prometheus.NewDesc("slurm_queue_timeout", "Jobs stopped by timeout", nil, nil),
-		preempted:     prometheus.NewDesc("slurm_queue_preempted", "Number of preempted jobs", nil, nil),
-		node_fail:     prometheus.NewDesc("slurm_queue_node_fail", "Number of jobs stopped due to node fail", nil, nil),
-		out_of_memory: prometheus.NewDesc("slurm_queue_out_of_memory", "Number of jobs stopped by oomkiller", nil, nil),
-	}
-}
-
 type QueueCollector struct {
-	pending       *prometheus.Desc
-	pending_dep   *prometheus.Desc
-	running       *prometheus.Desc
-	suspended     *prometheus.Desc
-	cancelled     *prometheus.Desc
-	completing    *prometheus.Desc
-	completed     *prometheus.Desc
-	configuring   *prometheus.Desc
-	failed        *prometheus.Desc
-	timeout       *prometheus.Desc
-	preempted     *prometheus.Desc
-	node_fail     *prometheus.Desc
-	out_of_memory *prometheus.Desc
+	pending     *prometheus.Desc
+	pendingDep  *prometheus.Desc
+	running     *prometheus.Desc
+	suspended   *prometheus.Desc
+	cancelled   *prometheus.Desc
+	completing  *prometheus.Desc
+	completed   *prometheus.Desc
+	configuring *prometheus.Desc
+	failed      *prometheus.Desc
+	timeout     *prometheus.Desc
+	preempted   *prometheus.Desc
+	nodeFail    *prometheus.Desc
+	outOfMemory *prometheus.Desc
+	logger      log.Logger
 }
 
-func (qc *QueueCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- qc.pending
-	ch <- qc.pending_dep
-	ch <- qc.running
-	ch <- qc.suspended
-	ch <- qc.cancelled
-	ch <- qc.completing
-	ch <- qc.completed
-	ch <- qc.configuring
-	ch <- qc.failed
-	ch <- qc.timeout
-	ch <- qc.preempted
-	ch <- qc.node_fail
-	ch <- qc.out_of_memory
+func init() {
+	registerCollector("queue", defaultEnabled, NewQueueCollector)
 }
 
-func (qc *QueueCollector) Collect(ch chan<- prometheus.Metric) {
-	qm := QueueGetMetrics()
+func NewQueueCollector(logger log.Logger) (Collector, error) {
+	return &QueueCollector{
+		logger:      logger,
+		pending:     prometheus.NewDesc("slurm_queue_pending", "Pending jobs in queue", nil, nil),
+		pendingDep:  prometheus.NewDesc("slurm_queue_pending_dependency", "Pending jobs because of dependency in queue", nil, nil),
+		running:     prometheus.NewDesc("slurm_queue_running", "Running jobs in the cluster", nil, nil),
+		suspended:   prometheus.NewDesc("slurm_queue_suspended", "Suspended jobs in the cluster", nil, nil),
+		cancelled:   prometheus.NewDesc("slurm_queue_cancelled", "Cancelled jobs in the cluster", nil, nil),
+		completing:  prometheus.NewDesc("slurm_queue_completing", "Completing jobs in the cluster", nil, nil),
+		completed:   prometheus.NewDesc("slurm_queue_completed", "Completed jobs in the cluster", nil, nil),
+		configuring: prometheus.NewDesc("slurm_queue_configuring", "Configuring jobs in the cluster", nil, nil),
+		failed:      prometheus.NewDesc("slurm_queue_failed", "Number of failed jobs", nil, nil),
+		timeout:     prometheus.NewDesc("slurm_queue_timeout", "Jobs stopped by timeout", nil, nil),
+		preempted:   prometheus.NewDesc("slurm_queue_preempted", "Number of preempted jobs", nil, nil),
+		nodeFail:    prometheus.NewDesc("slurm_queue_node_fail", "Number of jobs stopped due to node fail", nil, nil),
+		outOfMemory: prometheus.NewDesc("slurm_queue_out_of_memory", "Number of jobs stopped by oomkiller", nil, nil),
+	}, nil
+}
+
+func (qc *QueueCollector) Collect(ch chan<- prometheus.Metric) error {
+	out, err := RunCommand("squeue", "-a", "-r", "-h", "-o %A,%T,%r", "--states=all")
+	if err != nil {
+		return err
+	}
+
+	qm := ParseQueueMetrics(out)
 	ch <- prometheus.MustNewConstMetric(qc.pending, prometheus.GaugeValue, qm.pending)
-	ch <- prometheus.MustNewConstMetric(qc.pending_dep, prometheus.GaugeValue, qm.pending_dep)
+	ch <- prometheus.MustNewConstMetric(qc.pendingDep, prometheus.GaugeValue, qm.pendingDep)
 	ch <- prometheus.MustNewConstMetric(qc.running, prometheus.GaugeValue, qm.running)
 	ch <- prometheus.MustNewConstMetric(qc.suspended, prometheus.GaugeValue, qm.suspended)
 	ch <- prometheus.MustNewConstMetric(qc.cancelled, prometheus.GaugeValue, qm.cancelled)
@@ -171,6 +140,8 @@ func (qc *QueueCollector) Collect(ch chan<- prometheus.Metric) {
 	ch <- prometheus.MustNewConstMetric(qc.failed, prometheus.GaugeValue, qm.failed)
 	ch <- prometheus.MustNewConstMetric(qc.timeout, prometheus.GaugeValue, qm.timeout)
 	ch <- prometheus.MustNewConstMetric(qc.preempted, prometheus.GaugeValue, qm.preempted)
-	ch <- prometheus.MustNewConstMetric(qc.node_fail, prometheus.GaugeValue, qm.node_fail)
-	ch <- prometheus.MustNewConstMetric(qc.out_of_memory, prometheus.GaugeValue, qm.out_of_memory)
+	ch <- prometheus.MustNewConstMetric(qc.nodeFail, prometheus.GaugeValue, qm.nodeFail)
+	ch <- prometheus.MustNewConstMetric(qc.outOfMemory, prometheus.GaugeValue, qm.outOfMemory)
+
+	return nil
 }
