@@ -1,4 +1,6 @@
-/* Copyright 2017 Victor Penso, Matteo Dessalvi
+/*
+	Copyright 2017 Victor Penso, Matteo Dessalvi
+	Copyright 2024 Oleh Astappiev
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,58 +18,32 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package collector
 
 import (
-	"github.com/prometheus/client_golang/prometheus"
-	"io/ioutil"
-	"log"
-	"os/exec"
 	"regexp"
 	"strconv"
 	"strings"
+
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
-/*
- * Execute the Slurm sdiag command to read the current statistics
- * from the Slurm scheduler. It will be repreatedly called by the
- * collector.
- */
-
-// Basic metrics for the scheduler
 type SchedulerMetrics struct {
-	threads                           float64
-	queue_size                        float64
-	dbd_queue_size                    float64
-	last_cycle                        float64
-	mean_cycle                        float64
-	cycle_per_minute                  float64
-	backfill_last_cycle               float64
-	backfill_mean_cycle               float64
-	backfill_depth_mean               float64
-	total_backfilled_jobs_since_start float64
-	total_backfilled_jobs_since_cycle float64
-	total_backfilled_heterogeneous    float64
+	threads                       float64
+	queueSize                     float64
+	dbdQueueSize                  float64
+	lastCycle                     float64
+	meanCycle                     float64
+	cyclePerMinute                float64
+	backfillLastCycle             float64
+	backfillMeanCycle             float64
+	backfillDepthMean             float64
+	totalBackfilledJobsSinceStart float64
+	totalBackfilledJobsSinceCycle float64
+	totalBackfilledHeterogeneous  float64
 }
 
-// Execute the sdiag command and return its output
-func SchedulerData() []byte {
-	cmd := exec.Command("sdiag")
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := cmd.Start(); err != nil {
-		log.Fatal(err)
-	}
-	out, _ := ioutil.ReadAll(stdout)
-	if err := cmd.Wait(); err != nil {
-		log.Fatal(err)
-	}
-	return out
-}
-
-// Extract the relevant metrics from the sdiag output
 func ParseSchedulerMetrics(input []byte) *SchedulerMetrics {
 	var sm SchedulerMetrics
-	lines := strings.Split(string(input), "\n")
+	lines := SplitLines(input)
 	// Guard variables to check for string repetitions in the output of sdiag
 	// (two occurencies of the following strings: 'Last cycle', 'Mean cycle')
 	lc_count := 0
@@ -89,163 +65,98 @@ func ParseSchedulerMetrics(input []byte) *SchedulerMetrics {
 			case st.MatchString(state) == true:
 				sm.threads, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case qs.MatchString(state) == true:
-				sm.queue_size, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.queueSize, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case dbd.MatchString(state) == true:
-				sm.dbd_queue_size, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.dbdQueueSize, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case lc.MatchString(state) == true:
 				if lc_count == 0 {
-					sm.last_cycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+					sm.lastCycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 					lc_count = 1
 				}
 				if lc_count == 1 {
-					sm.backfill_last_cycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+					sm.backfillLastCycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 				}
 			case mc.MatchString(state) == true:
 				if mc_count == 0 {
-					sm.mean_cycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+					sm.meanCycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 					mc_count = 1
 				}
 				if mc_count == 1 {
-					sm.backfill_mean_cycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+					sm.backfillMeanCycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 				}
 			case cpm.MatchString(state) == true:
-				sm.cycle_per_minute, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.cyclePerMinute, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case dpm.MatchString(state) == true:
-				sm.backfill_depth_mean, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.backfillDepthMean, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case tbs.MatchString(state) == true:
-				sm.total_backfilled_jobs_since_start, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.totalBackfilledJobsSinceStart, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case tbc.MatchString(state) == true:
-				sm.total_backfilled_jobs_since_cycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.totalBackfilledJobsSinceCycle, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			case tbh.MatchString(state) == true:
-				sm.total_backfilled_heterogeneous, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
+				sm.totalBackfilledHeterogeneous, _ = strconv.ParseFloat(strings.TrimSpace(strings.Split(line, ":")[1]), 64)
 			}
 		}
 	}
 	return &sm
 }
 
-// Returns the scheduler metrics
-func SchedulerGetMetrics() *SchedulerMetrics {
-	return ParseSchedulerMetrics(SchedulerData())
-}
-
-/*
- * Implement the Prometheus Collector interface and feed the
- * Slurm scheduler metrics into it.
- * https://godoc.org/github.com/prometheus/client_golang/prometheus#Collector
- */
-
-// Collector strcture
 type SchedulerCollector struct {
-	threads                           *prometheus.Desc
-	queue_size                        *prometheus.Desc
-	dbd_queue_size                    *prometheus.Desc
-	last_cycle                        *prometheus.Desc
-	mean_cycle                        *prometheus.Desc
-	cycle_per_minute                  *prometheus.Desc
-	backfill_last_cycle               *prometheus.Desc
-	backfill_mean_cycle               *prometheus.Desc
-	backfill_depth_mean               *prometheus.Desc
-	total_backfilled_jobs_since_start *prometheus.Desc
-	total_backfilled_jobs_since_cycle *prometheus.Desc
-	total_backfilled_heterogeneous    *prometheus.Desc
+	threads                       *prometheus.Desc
+	queueSize                     *prometheus.Desc
+	dbdQueueSize                  *prometheus.Desc
+	lastCycle                     *prometheus.Desc
+	meanCycle                     *prometheus.Desc
+	cyclePerMinute                *prometheus.Desc
+	backfillLastCycle             *prometheus.Desc
+	backfillMeanCycle             *prometheus.Desc
+	backfillDepthMean             *prometheus.Desc
+	totalBackfilledJobsSinceStart *prometheus.Desc
+	totalBackfilledJobsSinceCycle *prometheus.Desc
+	totalBackfilledHeterogeneous  *prometheus.Desc
+	logger                        log.Logger
 }
 
-// Send all metric descriptions
-func (c *SchedulerCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- c.threads
-	ch <- c.queue_size
-	ch <- c.dbd_queue_size
-	ch <- c.last_cycle
-	ch <- c.mean_cycle
-	ch <- c.cycle_per_minute
-	ch <- c.backfill_last_cycle
-	ch <- c.backfill_mean_cycle
-	ch <- c.backfill_depth_mean
-	ch <- c.total_backfilled_jobs_since_start
-	ch <- c.total_backfilled_jobs_since_cycle
-	ch <- c.total_backfilled_heterogeneous
+func init() {
+	registerCollector("scheduler", defaultEnabled, NewSchedulerCollector)
 }
 
-// Send the values of all metrics
-func (sc *SchedulerCollector) Collect(ch chan<- prometheus.Metric) {
-	sm := SchedulerGetMetrics()
-	ch <- prometheus.MustNewConstMetric(sc.threads, prometheus.GaugeValue, sm.threads)
-	ch <- prometheus.MustNewConstMetric(sc.queue_size, prometheus.GaugeValue, sm.queue_size)
-	ch <- prometheus.MustNewConstMetric(sc.dbd_queue_size, prometheus.GaugeValue, sm.dbd_queue_size)
-	ch <- prometheus.MustNewConstMetric(sc.last_cycle, prometheus.GaugeValue, sm.last_cycle)
-	ch <- prometheus.MustNewConstMetric(sc.mean_cycle, prometheus.GaugeValue, sm.mean_cycle)
-	ch <- prometheus.MustNewConstMetric(sc.cycle_per_minute, prometheus.GaugeValue, sm.cycle_per_minute)
-	ch <- prometheus.MustNewConstMetric(sc.backfill_last_cycle, prometheus.GaugeValue, sm.backfill_last_cycle)
-	ch <- prometheus.MustNewConstMetric(sc.backfill_mean_cycle, prometheus.GaugeValue, sm.backfill_mean_cycle)
-	ch <- prometheus.MustNewConstMetric(sc.backfill_depth_mean, prometheus.GaugeValue, sm.backfill_depth_mean)
-	ch <- prometheus.MustNewConstMetric(sc.total_backfilled_jobs_since_start, prometheus.GaugeValue, sm.total_backfilled_jobs_since_start)
-	ch <- prometheus.MustNewConstMetric(sc.total_backfilled_jobs_since_cycle, prometheus.GaugeValue, sm.total_backfilled_jobs_since_cycle)
-	ch <- prometheus.MustNewConstMetric(sc.total_backfilled_heterogeneous, prometheus.GaugeValue, sm.total_backfilled_heterogeneous)
-}
-
-// Returns the Slurm scheduler collector, used to register with the prometheus client
-func NewSchedulerCollector() *SchedulerCollector {
+func NewSchedulerCollector(logger log.Logger) (Collector, error) {
 	return &SchedulerCollector{
-		threads: prometheus.NewDesc(
-			"slurm_scheduler_threads",
-			"Information provided by the Slurm sdiag command, number of scheduler threads ",
-			nil,
-			nil),
-		queue_size: prometheus.NewDesc(
-			"slurm_scheduler_queue_size",
-			"Information provided by the Slurm sdiag command, length of the scheduler queue",
-			nil,
-			nil),
-		dbd_queue_size: prometheus.NewDesc(
-			"slurm_scheduler_dbd_queue_size",
-			"Information provided by the Slurm sdiag command, length of the DBD agent queue",
-			nil,
-			nil),
-		last_cycle: prometheus.NewDesc(
-			"slurm_scheduler_last_cycle",
-			"Information provided by the Slurm sdiag command, scheduler last cycle time in (microseconds)",
-			nil,
-			nil),
-		mean_cycle: prometheus.NewDesc(
-			"slurm_scheduler_mean_cycle",
-			"Information provided by the Slurm sdiag command, scheduler mean cycle time in (microseconds)",
-			nil,
-			nil),
-		cycle_per_minute: prometheus.NewDesc(
-			"slurm_scheduler_cycle_per_minute",
-			"Information provided by the Slurm sdiag command, number scheduler cycles per minute",
-			nil,
-			nil),
-		backfill_last_cycle: prometheus.NewDesc(
-			"slurm_scheduler_backfill_last_cycle",
-			"Information provided by the Slurm sdiag command, scheduler backfill last cycle time in (microseconds)",
-			nil,
-			nil),
-		backfill_mean_cycle: prometheus.NewDesc(
-			"slurm_scheduler_backfill_mean_cycle",
-			"Information provided by the Slurm sdiag command, scheduler backfill mean cycle time in (microseconds)",
-			nil,
-			nil),
-		backfill_depth_mean: prometheus.NewDesc(
-			"slurm_scheduler_backfill_depth_mean",
-			"Information provided by the Slurm sdiag command, scheduler backfill mean depth",
-			nil,
-			nil),
-		total_backfilled_jobs_since_start: prometheus.NewDesc(
-			"slurm_scheduler_backfilled_jobs_since_start_total",
-			"Information provided by the Slurm sdiag command, number of jobs started thanks to backfilling since last slurm start",
-			nil,
-			nil),
-		total_backfilled_jobs_since_cycle: prometheus.NewDesc(
-			"slurm_scheduler_backfilled_jobs_since_cycle_total",
-			"Information provided by the Slurm sdiag command, number of jobs started thanks to backfilling since last time stats where reset",
-			nil,
-			nil),
-		total_backfilled_heterogeneous: prometheus.NewDesc(
-			"slurm_scheduler_backfilled_heterogeneous_total",
-			"Information provided by the Slurm sdiag command, number of heterogeneous job components started thanks to backfilling since last Slurm start",
-			nil,
-			nil),
+		logger:                        logger,
+		threads:                       prometheus.NewDesc("slurm_scheduler_threads", "Information provided by the Slurm sdiag command, number of scheduler threads ", nil, nil),
+		queueSize:                     prometheus.NewDesc("slurm_scheduler_queue_size", "Information provided by the Slurm sdiag command, length of the scheduler queue", nil, nil),
+		dbdQueueSize:                  prometheus.NewDesc("slurm_scheduler_dbd_queue_size", "Information provided by the Slurm sdiag command, length of the DBD agent queue", nil, nil),
+		lastCycle:                     prometheus.NewDesc("slurm_scheduler_last_cycle", "Information provided by the Slurm sdiag command, scheduler last cycle time in (microseconds)", nil, nil),
+		meanCycle:                     prometheus.NewDesc("slurm_scheduler_mean_cycle", "Information provided by the Slurm sdiag command, scheduler mean cycle time in (microseconds)", nil, nil),
+		cyclePerMinute:                prometheus.NewDesc("slurm_scheduler_cycle_per_minute", "Information provided by the Slurm sdiag command, number scheduler cycles per minute", nil, nil),
+		backfillLastCycle:             prometheus.NewDesc("slurm_scheduler_backfill_last_cycle", "Information provided by the Slurm sdiag command, scheduler backfill last cycle time in (microseconds)", nil, nil),
+		backfillMeanCycle:             prometheus.NewDesc("slurm_scheduler_backfill_mean_cycle", "Information provided by the Slurm sdiag command, scheduler backfill mean cycle time in (microseconds)", nil, nil),
+		backfillDepthMean:             prometheus.NewDesc("slurm_scheduler_backfill_depth_mean", "Information provided by the Slurm sdiag command, scheduler backfill mean depth", nil, nil),
+		totalBackfilledJobsSinceStart: prometheus.NewDesc("slurm_scheduler_backfilled_jobs_since_start_total", "Information provided by the Slurm sdiag command, number of jobs started thanks to backfilling since last slurm start", nil, nil),
+		totalBackfilledJobsSinceCycle: prometheus.NewDesc("slurm_scheduler_backfilled_jobs_since_cycle_total", "Information provided by the Slurm sdiag command, number of jobs started thanks to backfilling since last time stats where reset", nil, nil),
+		totalBackfilledHeterogeneous:  prometheus.NewDesc("slurm_scheduler_backfilled_heterogeneous_total", "Information provided by the Slurm sdiag command, number of heterogeneous job components started thanks to backfilling since last Slurm start", nil, nil),
+	}, nil
+}
+
+func (sc *SchedulerCollector) Collect(ch chan<- prometheus.Metric) error {
+	out, err := RunCommand("sdiag")
+	if err != nil {
+		return err
 	}
+
+	sm := ParseSchedulerMetrics(out)
+	ch <- prometheus.MustNewConstMetric(sc.threads, prometheus.GaugeValue, sm.threads)
+	ch <- prometheus.MustNewConstMetric(sc.queueSize, prometheus.GaugeValue, sm.queueSize)
+	ch <- prometheus.MustNewConstMetric(sc.dbdQueueSize, prometheus.GaugeValue, sm.dbdQueueSize)
+	ch <- prometheus.MustNewConstMetric(sc.lastCycle, prometheus.GaugeValue, sm.lastCycle)
+	ch <- prometheus.MustNewConstMetric(sc.meanCycle, prometheus.GaugeValue, sm.meanCycle)
+	ch <- prometheus.MustNewConstMetric(sc.cyclePerMinute, prometheus.GaugeValue, sm.cyclePerMinute)
+	ch <- prometheus.MustNewConstMetric(sc.backfillLastCycle, prometheus.GaugeValue, sm.backfillLastCycle)
+	ch <- prometheus.MustNewConstMetric(sc.backfillMeanCycle, prometheus.GaugeValue, sm.backfillMeanCycle)
+	ch <- prometheus.MustNewConstMetric(sc.backfillDepthMean, prometheus.GaugeValue, sm.backfillDepthMean)
+	ch <- prometheus.MustNewConstMetric(sc.totalBackfilledJobsSinceStart, prometheus.GaugeValue, sm.totalBackfilledJobsSinceStart)
+	ch <- prometheus.MustNewConstMetric(sc.totalBackfilledJobsSinceCycle, prometheus.GaugeValue, sm.totalBackfilledJobsSinceCycle)
+	ch <- prometheus.MustNewConstMetric(sc.totalBackfilledHeterogeneous, prometheus.GaugeValue, sm.totalBackfilledHeterogeneous)
+
+	return nil
 }
