@@ -1,4 +1,6 @@
-/* Copyright 2021 Victor Penso
+/*
+	Copyright 2021 Victor Penso
+	Copyright 2024 Oleh Astappiev
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,46 +18,30 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package collector
 
 import (
-	"github.com/prometheus/client_golang/prometheus"
-	"io/ioutil"
-	"log"
-	"os/exec"
 	"strconv"
 	"strings"
-)
 
-func FairShareData() []byte {
-	cmd := exec.Command("sshare", "-n", "-P", "-o", "account,fairshare")
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := cmd.Start(); err != nil {
-		log.Fatal(err)
-	}
-	out, _ := ioutil.ReadAll(stdout)
-	if err := cmd.Wait(); err != nil {
-		log.Fatal(err)
-	}
-	return out
-}
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
+)
 
 type FairShareMetrics struct {
 	fairshare float64
 }
 
-func ParseFairShareMetrics() map[string]*FairShareMetrics {
+func ParseFairShareMetrics(input []byte) map[string]*FairShareMetrics {
 	accounts := make(map[string]*FairShareMetrics)
-	lines := strings.Split(string(FairShareData()), "\n")
-	for _, line := range lines {
+
+	for _, line := range SplitLines(input) {
 		if !strings.HasPrefix(line, "  ") {
 			if strings.Contains(line, "|") {
-				account := strings.Trim(strings.Split(line, "|")[0], " ")
+				parts := strings.Split(line, "|")
+				account := strings.Trim(parts[0], " ")
 				_, key := accounts[account]
 				if !key {
-					accounts[account] = &FairShareMetrics{0}
+					accounts[account] = &FairShareMetrics{}
 				}
-				fairshare, _ := strconv.ParseFloat(strings.Split(line, "|")[1], 64)
+				fairshare, _ := strconv.ParseFloat(parts[1], 64)
 				accounts[account].fairshare = fairshare
 			}
 		}
@@ -65,22 +51,30 @@ func ParseFairShareMetrics() map[string]*FairShareMetrics {
 
 type FairShareCollector struct {
 	fairshare *prometheus.Desc
+	logger    log.Logger
 }
 
-func NewFairShareCollector() *FairShareCollector {
-	labels := []string{"account"}
+func init() {
+	registerCollector("fairshare", defaultDisabled, NewFairShareCollector)
+}
+
+func NewFairShareCollector(logger log.Logger) (Collector, error) {
 	return &FairShareCollector{
-		fairshare: prometheus.NewDesc("slurm_account_fairshare", "FairShare for account", labels, nil),
+		logger:    logger,
+		fairshare: prometheus.NewDesc("slurm_account_fairshare", "FairShare for account", []string{"account"}, nil),
+	}, nil
+}
+
+func (fsc *FairShareCollector) Collect(ch chan<- prometheus.Metric) error {
+	out, err := RunCommand("sshare", "-n", "-P", "-o", "account,fairshare")
+	if err != nil {
+		return err
 	}
-}
 
-func (fsc *FairShareCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- fsc.fairshare
-}
-
-func (fsc *FairShareCollector) Collect(ch chan<- prometheus.Metric) {
-	fsm := ParseFairShareMetrics()
+	fsm := ParseFairShareMetrics(out)
 	for f := range fsm {
 		ch <- prometheus.MustNewConstMetric(fsc.fairshare, prometheus.GaugeValue, fsm[f].fairshare, f)
 	}
+
+	return nil
 }
