@@ -16,52 +16,187 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package main
 
 import (
-	"flag"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/vpenso/prometheus-slurm-exporter/collector"
-	"log"
+	"fmt"
+	stdlog "log"
 	"net/http"
+	"os"
+	"sort"
+
+	"github.com/alecthomas/kingpin/v2"
+	"github.com/go-kit/kit/log/level"
+	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
+	promcollectors "github.com/prometheus/client_golang/prometheus/collectors"
+	versioncollector "github.com/prometheus/client_golang/prometheus/collectors/version"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/prometheus/common/promlog"
+	"github.com/prometheus/common/promlog/flag"
+	"github.com/prometheus/common/version"
+	"github.com/prometheus/exporter-toolkit/web"
+	"github.com/prometheus/exporter-toolkit/web/kingpinflag"
+
+	"github.com/vpenso/prometheus-slurm-exporter/collector"
 )
 
-func init() {
-	// Metrics have to be registered to be exposed
-	prometheus.MustRegister(collector.NewAccountsCollector())   // from accounts.go
-	prometheus.MustRegister(collector.NewCPUsCollector())       // from cpus.go
-	prometheus.MustRegister(collector.NewNodesCollector())      // from nodes.go
-	prometheus.MustRegister(collector.NewNodeCollector())       // from node.go
-	prometheus.MustRegister(collector.NewPartitionsCollector()) // from partitions.go
-	prometheus.MustRegister(collector.NewQueueCollector())      // from queue.go
-	prometheus.MustRegister(collector.NewSchedulerCollector())  // from scheduler.go
-	prometheus.MustRegister(collector.NewFairShareCollector())  // from sshare.go
-	prometheus.MustRegister(collector.NewUsersCollector())      // from users.go
-	prometheus.MustRegister(collector.NewJobsCollector())       // from jobs.go
+var (
+	metricsPath            = kingpin.Flag("web.telemetry-path", "Path under which to expose metrics.").Default("/metrics").String()
+	disableExporterMetrics = kingpin.Flag("web.disable-exporter-metrics", "Exclude metrics about the exporter itself (promhttp_*, process_*, go_*).").Bool()
+	maxRequests            = kingpin.Flag("web.max-requests", "Maximum number of parallel scrape requests. Use 0 to disable.").Default("40").Int()
+	toolkitFlags           = kingpinflag.AddFlags(kingpin.CommandLine, ":9341")
+)
 
-	log.Printf("Initializing main program")
+// handler wraps an unfiltered http.Handler but uses a filtered handler,
+// created on the fly, if filtering is requested. Create instances with
+// newHandler.
+type handler struct {
+	unfilteredHandler http.Handler
+	// exporterMetricsRegistry is a separate registry for the metrics about
+	// the exporter itself.
+	exporterMetricsRegistry *prometheus.Registry
+	includeExporterMetrics  bool
+	maxRequests             int
+	logger                  log.Logger
 }
 
-var listenAddress = flag.String(
-	"listen-address",
-	":8080",
-	"The address to listen on for HTTP requests.")
+// ServeHTTP implements http.Handler.
+func (h *handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	filters := r.URL.Query()["collect[]"]
+	level.Debug(h.logger).Log("msg", "collect query:", "filters", filters)
 
-var gpuAcct = flag.Bool(
-	"gpus-acct",
-	false,
-	"Enable GPUs accounting")
+	if len(filters) == 0 {
+		// No filters, use the prepared unfiltered handler.
+		h.unfilteredHandler.ServeHTTP(w, r)
+		return
+	}
+	// To serve filtered metrics, we create a filtering handler on the fly.
+	filteredHandler, err := h.innerHandler(filters...)
+	if err != nil {
+		level.Warn(h.logger).Log("msg", "Couldn't create filtered metrics handler:", "err", err)
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte(fmt.Sprintf("Couldn't create filtered metrics handler: %s", err)))
+		return
+	}
+	filteredHandler.ServeHTTP(w, r)
+}
 
-func main() {
-	flag.Parse()
+func newHandler(includeExporterMetrics bool, maxRequests int, logger log.Logger) *handler {
+	h := &handler{
+		exporterMetricsRegistry: prometheus.NewRegistry(),
+		includeExporterMetrics:  includeExporterMetrics,
+		maxRequests:             maxRequests,
+		logger:                  logger,
+	}
+	if h.includeExporterMetrics {
+		h.exporterMetricsRegistry.MustRegister(
+			promcollectors.NewProcessCollector(promcollectors.ProcessCollectorOpts{}),
+			promcollectors.NewGoCollector(),
+		)
+	}
+	if innerHandler, err := h.innerHandler(); err != nil {
+		panic(fmt.Sprintf("Couldn't create metrics handler: %s", err))
+	} else {
+		h.unfilteredHandler = innerHandler
+	}
+	return h
+}
 
-	// Turn on GPUs accounting only if the corresponding command line option is set to true.
-	if *gpuAcct {
-		prometheus.MustRegister(collector.NewGPUsCollector()) // from gpus.go
+// innerHandler is used to create both the one unfiltered http.Handler to be
+// wrapped by the outer handler and also the filtered handlers created on the
+// fly. The former is accomplished by calling innerHandler without any arguments
+// (in which case it will log all the collectors enabled via command-line
+// flags).
+func (h *handler) innerHandler(filters ...string) (http.Handler, error) {
+	nc, err := collector.NewSlurmCollector(h.logger, filters...)
+	if err != nil {
+		return nil, fmt.Errorf("couldn't create collector: %s", err)
 	}
 
-	// The Handler function provides a default handler to expose metrics
-	// via an HTTP server. "/metrics" is the usual endpoint for that.
-	log.Printf("Starting Server: %s", *listenAddress)
-	log.Printf("GPUs Accounting: %t", *gpuAcct)
-	http.Handle("/metrics", promhttp.Handler())
-	log.Fatal(http.ListenAndServe(*listenAddress, nil))
+	// Only log the creation of an unfiltered handler, which should happen
+	// only once upon startup.
+	if len(filters) == 0 {
+		level.Info(h.logger).Log("msg", "Enabled collectors")
+		collectors := []string{}
+		for n := range nc.Collectors {
+			collectors = append(collectors, n)
+		}
+		sort.Strings(collectors)
+		for _, c := range collectors {
+			level.Info(h.logger).Log("collector", c)
+		}
+	}
+
+	r := prometheus.NewRegistry()
+	r.MustRegister(versioncollector.NewCollector("slurm_exporter"))
+	if err := r.Register(nc); err != nil {
+		return nil, fmt.Errorf("couldn't register node collector: %s", err)
+	}
+
+	var handler http.Handler
+	if h.includeExporterMetrics {
+		handler = promhttp.HandlerFor(
+			prometheus.Gatherers{h.exporterMetricsRegistry, r},
+			promhttp.HandlerOpts{
+				ErrorLog:            stdlog.New(log.NewStdlibAdapter(level.Error(h.logger)), "", 0),
+				ErrorHandling:       promhttp.ContinueOnError,
+				MaxRequestsInFlight: h.maxRequests,
+				Registry:            h.exporterMetricsRegistry,
+			},
+		)
+		// Note that we have to use h.exporterMetricsRegistry here to
+		// use the same promhttp metrics for all expositions.
+		handler = promhttp.InstrumentMetricHandler(
+			h.exporterMetricsRegistry, handler,
+		)
+	} else {
+		handler = promhttp.HandlerFor(
+			r,
+			promhttp.HandlerOpts{
+				ErrorLog:            stdlog.New(log.NewStdlibAdapter(level.Error(h.logger)), "", 0),
+				ErrorHandling:       promhttp.ContinueOnError,
+				MaxRequestsInFlight: h.maxRequests,
+			},
+		)
+	}
+
+	return handler, nil
+}
+
+func main() {
+	promlogConfig := &promlog.Config{}
+	flag.AddFlags(kingpin.CommandLine, promlogConfig)
+	kingpin.Version(version.Print("slurm_exporter"))
+	kingpin.CommandLine.UsageWriter(os.Stdout)
+	kingpin.HelpFlag.Short('h')
+	kingpin.Parse()
+	logger := promlog.New(promlogConfig)
+
+	level.Info(logger).Log("msg", "Starting slurm_exporter", "version", version.Info())
+	level.Info(logger).Log("msg", "Build context", "context", version.BuildContext())
+
+	http.Handle(*metricsPath, newHandler(!*disableExporterMetrics, *maxRequests, logger))
+	if *metricsPath != "/" && *metricsPath != "" {
+		landingConfig := web.LandingConfig{
+			Name:        "SLURM Exporter",
+			Description: "Prometheus Exporter for Slurm Workload Manager",
+			Version:     version.Info(),
+			Links: []web.LandingLinks{
+				{
+					Address: *metricsPath,
+					Text:    "Metrics",
+				},
+			},
+		}
+		landingPage, err := web.NewLandingPage(landingConfig)
+		if err != nil {
+			level.Error(logger).Log("err", err)
+			os.Exit(1)
+		}
+		http.Handle("/", landingPage)
+	}
+
+	server := &http.Server{}
+	if err := web.ListenAndServe(server, toolkitFlags, logger); err != nil {
+		level.Error(logger).Log("msg", "Error starting HTTP server", "err", err)
+		os.Exit(1)
+	}
 }
